fix android
